Store number type as f32 instead of i32.

+++ src/ast.rs
@@ -15,7 +15,7 @@ pub enum Factor {
 
 pub enum Term {
     String(String),
-    Num(i32),
+    Num(f32),
 }
 
 pub enum OpKind {
